chore: move /stake/btc/details to /stake/btc
Currently /stake/btc is the finality providers page for btc staking. However, this route should contain the UI that's currently on /stake/btc/details.

- Move /stake/btc to /stake/btc/providers
- Move /stake/btc/details to /stake/btc
