+++ src/app/stake/btc/providers/MobileFinalityProvidersView.tsx
